compatible with past committee of subchain for a certain time

+++ contracts/ultrainio.system/voting.cpp
@@ -41,8 +41,8 @@ namespace ultrainiosystem {
       require_auth( producer );
 
       auto prod = _producers.find( producer );
-      uint64_t curblocknum = tapos_block_num();
-      if ( prod != _producers.end() ) {
+      uint64_t curblocknum = (uint64_t)tapos_block_num();
+      if( prod != _producers.end() ) {
          ultrainio_assert( (curblocknum - prod->last_operate_blocknum) > 2 , "interval operate at least more than certain number block high" );
          //if location changes
          if(prod->is_enabled && prod->location != location) {
@@ -72,7 +72,7 @@ namespace ultrainiosystem {
                info.url          = url;
                info.location     = location;
                info.last_operate_blocknum = curblocknum;
-            });
+         });
       } else {
          _producers.emplace( producer, [&]( producer_info& info ){
                info.owner         = producer;
@@ -91,7 +91,7 @@ namespace ultrainiosystem {
       require_auth( producer );
 
       const auto& prod = _producers.get( producer, "producer not found" );
-      uint64_t curblocknum = tapos_block_num();
+      uint64_t curblocknum = (uint64_t)tapos_block_num();
       print("unregprod tapos_block_num:",curblocknum," prod->last_operate_blocknum:",prod.last_operate_blocknum);
 
       ultrainio_assert( (curblocknum - prod.last_operate_blocknum) > 2 , "interval operate at least more than certain number block high" );
